# ReciFEED/api-docs/monolith/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ReciFEED API - Monolith",
    "description": "The monolith section of a hybrid microservices-monolith API forming the ReciFEED website app's backend.",
    "version": "1.0.0",
    "servers": [
      {
        "url": "http://recifeed.example.com",
        "description": "Recifeed API URL from Ingress"
      }
    ]
  },
  "paths": {
...
    "/api/users/login": {
      "post": {
        "tags": ["Users"],
        "summary": "Authenticates a user with email and password, returns JWT token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "User email address"
                  },
                  "password": {
                    "type": "string",
                    "description": "User password"
                  }
                }
              },
              "example": {
                "email": "john@gmail.com",
                "password": "password123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, returns JWT token and user info.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "token": {
                      "type": "string",
                      "description": "JWT bearer token"
                    },
                    "id": {
                      "type": "string",
                      "description": "User ID"
                    },
                    "username": {
                      "type": "string",
                      "description": "Username"
                    }
                  }
                },
                "example": {
                  "message": "Login successful",
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY5MzQ3OWE2YjRkODNiYTc1OTdiMTBjOSIsInVzZXJuYW1lIjoiSm9obkRvZSIsImlhdCI6MTYzMzA1NjAwMCwiZXhwIjoxNjMzMDk5MjAwfQ.abc123xyz",
                  "id": "693479a6b4d83ba7597b10c9",
                  "username": "JohnDoe"
                }
              }
            }
          },
          "400": {
            "description": "Missing email or password.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "status": "error",
                  "statusCode": 400,
                  "message": "Email and password required for login."
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "status": "error",
                  "statusCode": 401,
                  "message": "Invalid email or password."
                }
              }
            }
          },
          "500": { "$ref": "#/components/schemas/500Error" }
        }
      }
    },
    "/api/users/validate": {
      "get": {
        "tags": ["Users"],
        "summary": "Validates a JWT token and returns decoded user information.",
        "requestBody": {
          "required": false,
          "description": "Token is passed via Authorization header, not request body."
        },
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Token is valid, returns decoded user data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "User ID from token"
                        },
                        "username": {
                          "type": "string",
                          "description": "Username from token"
                        },
                        "iat": {
                          "type": "number",
                          "description": "Token issued at timestamp"
                        },
                        "exp": {
                          "type": "number",
                          "description": "Token expiration timestamp"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "valid": true,
                  "user": {
                    "id": "693479a6b4d83ba7597b10c9",
                    "username": "JohnDoe",
                    "iat": 1633056000,
                    "exp": 1633099200
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "status": "error",
                  "statusCode": 401,
                  "message": "Unauthorized: Missing or invalid token."
                }
              }
            }
          },
          "500": { "$ref": "#/components/schemas/500Error" }
        }
      }
    },
    "/api/users": {
      "post": {
        "tags": ["Users"],
        "summary": "Creates a new user account.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "username"],
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "User email address"
                  },
                  "password": {
                    "type": "string",
                    "description": "User password"
                  },
                  "username": {
                    "type": "string",
                    "description": "Username"
                  }
                }
              },
              "example": {
                "email": "newuser@example.com",
                "password": "securepassword",
                "username": "NewUser123"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "username": {
                          "type": "string"
                        },
                        "__v": {
                          "type": "number"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "status": "success",
                  "data": {
                    "_id": "69348446b4d83ba7597b11c3",
                    "created_at": "2025-12-07T01:15:00.000Z",
                    "email": "newuser@example.com",
                    "username": "NewUser123",
                    "__v": 0
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields or user already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "missingFields": {
                    "value": {
                      "status": "error",
                      "statusCode": 400,
                      "message": "Please provide email, password, and username"
                    }
                  },
                  "userExists": {
                    "value": {
                      "status": "error",
                      "statusCode": 400,
                      "message": "User with this email already exists"
                    }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/schemas/500Error" }
        }
      }
    },
    "/api/search": {
      "get": {
        "tags": ["Search"],
        "summary": "Searches across all entities (recipes, users, posts) with a single query.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Search query string (case-insensitive)",
            "schema": {
              "type": "string"
            },
            "example": "rice"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns search results from all entities.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "query": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "recipes": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "description": "See recipe endpoint schemas."
                          }
                        },
                        "users": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "description": "See user endpoint schemas."
                          }
                        },
                        "posts": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "description": "See post endpoint schemas."
                          }
                        },
                        "totalResults": {
                          "type": "number"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "status": "success",
                  "query": "pasta",
                  "data": {
                    "recipes": [
                      {
                        "_id": "692ff193a5480ded0a6fc658",
                        "created_at": "2025-11-01T18:20:00.000Z",
                        "user_id": "693278ef0ce44526e02f9290",
                        "username": "grace",
                        "title": "Butter Pasta",
                        "tags": ["Lunch", "Dinner"],
                        "ingredients": [
                          "1 cup noodles",
                          "2 cups water",
                          "1 tbsp butter"
                        ],
                        "instructions": [
                          "Boil water.",
                          "Add noodles and butter.",
                          "Simmer until done."
                        ],
                        "image_urls": [
                          "https://andianne.com/wp-content/uploads/2023/05/garlic-butter-noodles-02.jpg"
                        ],
                        "cooking_time": "25mins",
                        "reviews": [
                          {
                            "_id": "6934dff0cda8a8bfe83f88b8",
                            "user_id": "690f704d3fe0dd94aba6c975",
                            "username": "dave",
                            "rating": 4,
                            "comment": "Simple and satisfying!",
                            "created_at": "2025-11-01T18:45:00.000Z"
                          }
                        ]
                      }
                    ],
                    "users": [
                      {
                        "_id": "693479a6b4d83ba7597b10c9",
                        "created_at": "2025-10-28T16:20:00.000Z",
                        "username": "PastaLover",
                        "email": "pasta@example.com"
                      }
                    ],
                    "posts": [
                      {
                        "_id": "693479a6b4d83ba7597b10e5",
                        "created_at": "2025-11-04T17:30:00.000Z",
                        "user_id": "693479a6b4d83ba7597b10c9",
                        "username": "ivan",
                        "body": "Just made the best pasta dish!",
                        "image_urls": [
                          "https://cdn.apartmenttherapy.info/image/upload/f_jpg,q_auto:eco,c_fill,g_auto,w_1500,ar_1:1/k%2Farchive%2F2ba8aab83d28bb172962e3064b9da28047d8f49b",
                          "https://themenumaid.com/wp-content/uploads/2021/08/IMG_E1797-blog-photo.jpg"
                        ],
                        "likes": [],
                        "comments": [],
                        "__v": 2
                      }
                    ],
                    "totalResults": 3
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing search query parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "status": "error",
                  "statusCode": 400,
                  "message": "Please provide a search query"
                }
              }
            }
          },
          "401": { "$ref": "#/components/schemas/401TokenError" },
          "500": { "$ref": "#/components/schemas/500Error" }
        }
      }
    },
    "/api/search/recipes": {
      "get": {
        "tags": ["Search"],
        "summary": "Searches recipes by query, tags, and/or cooking time. All parameters optional.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "description": "Search query string for title, tags, or ingredients (case-insensitive)",
            "schema": {
              "type": "string"
            },
            "example": "toast"
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "description": "Comma-separated list of tags to filter by",
            "schema": {
              "type": "string"
            },
            "example": "breakfast"
          },
          {
            "name": "cookingTime",
            "in": "query",
            "required": false,
            "description": "Cooking time range filter",
            "schema": {
              "type": "string",
              "enum": ["0-15", "15-30", "30-45", "45-60", "60+"]
            },
            "example": "0-15"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns filtered recipe results (up to 50).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "query": {
                      "type": "string"
                    },
                    "filters": {
                      "type": "object",
                      "properties": {
                        "tags": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "cookingTime": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "recipes": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "totalResults": {
                          "type": "number"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "status": "success",
                  "query": "toast",
                  "filters": {
                    "tags": ["breakfast"],
                    "cookingTime": "0-15"
                  },
                  "data": {
                    "recipes": [
                      {
                        "_id": "692ff193a5480ded0a6fc65b",
                        "created_at": "2025-10-29T20:25:00.000Z",
                        "user_id": "693278ef0ce44526e02f9293",
                        "username": "judy",
                        "title": "Avocado Toast",
                        "tags": ["Breakfast", "Vegan", "Snack"],
                        "ingredients": [
                          "1 avocado",
                          "2 slices bread",
                          "Salt and pepper"
                        ],
                        "instructions": [
                          "Mash avocado with seasoning.",
                          "Spread on toasted bread."
                        ],
                        "image_urls": [
                          "https://thesaltypot.com/wp-content/uploads/2019/02/Avocado-Toast09.jpg"
                        ],
                        "cooking_time": "10mins",
                        "reviews": [
                          {
                            "_id": "6934e2aecda8a8bfe83f88c7",
                            "user_id": "690f704d3fe0dd94aba6c973",
                            "username": "bob",
                            "rating": 5,
                            "comment": "Simple and delicious!",
                            "created_at": "2025-10-29T21:30:00.000Z"
                          },
                          {
                            "_id": "6934e2aecda8a8bfe83f88c8",
                            "user_id": "690f704d3fe0dd94aba6c973",
                            "username": "bob",
                            "rating": 5,
                            "comment": "Again, great recipe!",
                            "created_at": "2025-10-29T21:40:00.000Z"
                          }
                        ]
                      }
                    ],
                    "totalResults": 1
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/schemas/401TokenError" },
          "500": { "$ref": "#/components/schemas/500Error" }
        }
      }
    },
    "/api/search/users": {
      "get": {
        "tags": ["Search"],
        "summary": "Searches users by username or email.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Search query string for username or email (case-insensitive)",
            "schema": {
              "type": "string"
            },
            "example": "john"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns matching users (up to 50, sans passwords).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "query": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "users": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "_id": {
                                "type": "string"
                              },
                              "created_at": {
                                "type": "string"
                              },
                              "email": {
                                "type": "string"
                              },
                              "username": {
                                "type": "string"
                              },
                              "__v": {
                                "type": "number"
                              }
                            }
                          }
                        },
                        "totalResults": {
                          "type": "number"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "status": "success",
                  "query": "john",
                  "data": {
                    "users": [
                      {
                        "_id": "693479a6b4d83ba7597b10c9",
                        "created_at": "2025-12-06T18:44:54.591Z",
                        "email": "john@gmail.com",
                        "username": "JohnDoe",
                        "__v": 0
                      }
                    ],
                    "totalResults": 1
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing search query parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "status": "error",
                  "statusCode": 400,
                  "message": "Please provide a search query"
                }
              }
            }
          },
          "401": { "$ref": "#/components/schemas/401TokenError" },
          "500": { "$ref": "#/components/schemas/500Error" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "500Error": {
        "description": "Internal error, returns appropriate error response.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string"
                },
                "statusCode": {
                  "type": "number"
                },
                "message": {
                  "type": "string"
                }
              }
            },
            "example": {
              "status": "error",
              "statusCode": 500,
              "message": "Internal server error"
            }
          }
        }
      },
      "401TokenError": {
        "description": "Authentication error, no valid bearer token.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string"
                },
                "statusCode": {
                  "type": "number"
                },
                "message": {
                  "type": "string"
                }
              }
            },
            "example": {
              "status": "error",
              "statusCode": 500,
              "message": "Internal server error"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Required for most all endpoints in this API (sans login, user create, user validate, checks). Received from a valid ```/login``` response."
      }
    }
  }
}